+++ src/sonzai/resources/__init__.py
@@ -23,6 +23,8 @@ from .org import AsyncOrg, Org
 from .pipelines import AsyncPipelines, Pipelines
 from .project_config import AsyncProjectConfig, ProjectConfig
 from .project_notifications import AsyncProjectNotifications, ProjectNotifications
+from .routing import AsyncRouting, Routing
+from .runtime import AsyncRuntime, Runtime
 from .schedules import AsyncSchedules, Schedules
 from .skills import AsyncSkills, Skills
 from .storefront import AsyncStorefront, Storefront
@@ -76,6 +78,10 @@ __all__ = [
     "AsyncProjectConfig",
     "ProjectNotifications",
     "AsyncProjectNotifications",
+    "Runtime",
+    "AsyncRuntime",
+    "Routing",
+    "AsyncRouting",
     "Schedules",
     "AsyncSchedules",
     "Skills",